feat(...) : repo 추가 및 insert 변경
