+++ app/src/main/java/com/iron/espresso/domain/repo/ChatRepository.kt
@@ -1,5 +1,6 @@
 package com.iron.espresso.domain.repo
 
+import com.google.gson.JsonObject
 import com.iron.espresso.local.model.ChatEntity
 import io.reactivex.Completable
 import io.reactivex.Flowable
@@ -10,4 +11,10 @@ interface ChatRepository {
     fun insert(chatEntity: ChatEntity): Completable
 
     fun setChat(studyId: Int, first: Boolean)
+
+    fun onConnect(studyId: Int)
+
+    fun onDisconnect()
+
+    fun sendMessage(chatMessage: JsonObject)
 }
