+++ app/src/main/java/com/iron/espresso/presentation/home/mystudy/studydetail/chat/ChattingViewModel.kt
@@ -9,20 +9,14 @@ import com.google.gson.JsonObject
 import com.iron.espresso.Logger
 import com.iron.espresso.base.BaseViewModel
 import com.iron.espresso.domain.repo.ChatRepository
-import com.iron.espresso.ext.networkSchedulers
 import com.iron.espresso.ext.plusAssign
-import com.iron.espresso.local.model.ChatEntity
-import com.iron.espresso.model.source.remote.ChatRemoteDataSource
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import io.socket.emitter.Emitter
-import org.json.JSONObject
 
 class ChattingViewModel @ViewModelInject constructor(
     @Assisted
     private val savedState: SavedStateHandle,
-    private val chatRepository: ChatRepository,
-    private val chatRemoteDataSource: ChatRemoteDataSource
+    private val chatRepository: ChatRepository
 ) : BaseViewModel() {
 
     private val studyId: Int by lazy {
@@ -34,16 +28,6 @@ class ChattingViewModel @ViewModelInject constructor(
 
     private val _userNickname = MutableLiveData<String>()
     val userNickname: LiveData<String> get() = _userNickname
-
-    private val onChatReceiver = Emitter.Listener { args ->
-        val response = JSONObject(args.getOrNull(0).toString())
-
-        Logger.d("chatResponse: $response")
-    }
-
-    private val onChatSendReceiver = Emitter.Listener { args ->
-        Logger.d("${args.map { it.toString() }}")
-    }
 
     fun setChat(studyId: Int) {
         chatRepository.setChat(studyId, false)
@@ -61,22 +45,16 @@ class ChattingViewModel @ViewModelInject constructor(
             })
     }
 
-    fun insert(chatEntity: ChatEntity) {
-        compositeDisposable += chatRepository.insert(chatEntity)
-            .networkSchedulers()
-            .subscribe()
-    }
-
     fun onConnect() {
-        chatRemoteDataSource.onConnect(studyId, onChatSendReceiver, onChatReceiver)
+        chatRepository.onConnect(studyId)
     }
 
     fun onDisconnect() {
-        chatRemoteDataSource.onDisconnect(onChatSendReceiver, onChatReceiver)
+        chatRepository.onDisconnect()
     }
 
     fun sendMessage(message: JsonObject) {
-        chatRemoteDataSource.sendMessage(message)
+        chatRepository.sendMessage(message)
     }
 
     companion object {
